+++ sensoff/__init__.py
@@ -1,2 +1,2 @@
 from .__version__ import __version__
-from .sensoff import sensor_coordinates
+from .sensoff import Survey