+++ sensoff/sensoff.py
@@ -4,12 +4,7 @@ Corrects GPS transect coordinates when an on-the-go sensor is
 significantly offset from the GPS on a mobile instrument platform.
 
 """
-import argparse
-import csv
-from itertools import tee
-from math import atan2, cos, hypot, isclose, pi, sin, sqrt
-import sys
-
+from __future__ import annotations
 
 schematic = """
              lateral_offset
@@ -27,150 +22,50 @@ schematic = """
                   (-)
 
 """
+import argparse
+import csv
+from itertools import tee
+from math import atan2, cos, hypot, isclose, pi, sin, sqrt
+from typing import NamedTuple, List, Tuple, Iterable, Iterator, Any
+import sys
+
+# Standard Position Angle (SPA) == standard trigonometric unit circle
+# angle having one ray coincident with the positive x-axis, positive
+# in the counter-clockwise direction, and a value between (-pi, pi).
+StandardPositionAngle = float
+Point = Tuple[float, float]
 
 
-def sensor_coordinates(
-    gps_file,
-    inline_offset=0,
-    lateral_offset=0,
-    xcol=1,
-    ycol=2,
-    delimiter=",",
-    headrows=None,
-    ioff=None,
-    loff=None,
-):
-    """Convert GPS coordinates to offset sensor coordinates.
-    
-    Corrects GPS coordinates when on-the-go transect data have been
-    collected using a sensor that is significantly offset from the GPS
-    on a mobile platform.
-    
-    Parameters
-    ----------
-    gps_file : csv filename or object
-        Delimited file containing GPS coordinates for survey transect.
-    inline_offset : float, optional
-        Sensor offset distance inline with the direction of travel.
-        Positive in the direction of travel. Default is 0.
-    lateral_offset : float, optional
-        Sensor offset distance lateral to the direction of travel.
-        Positive to the left (facing forward). Default is 0.
-    xcol : int, optional
-        Column in datafile with x-coordinates. Default is 1.
-    ycol : int, optional
-        Column in datafile with y-coordinates. Default is 2.
-    delimiter : str, optional
-        Delimiter used in coordinate datafile. Default is ",".
-    headrows : int, optional
-        Number of header rows in datafile. Default is 0.
-    ioff, loff : floats, optional
-        Aliases for inline_offset and lateral_offset.
-
-    Returns
-    -------
-    List of original and adjusted x,y coordinates
-
-    """
-    if ioff is None:
-        ioff = inline_offset
-    if loff is None:
-        loff = lateral_offset
-
-    gps_coords = read_xy(gps_file, delimiter, xcol, ycol, headrows)
-    headings = platform_headings(gps_coords)
-    return offset_coords(gps_coords, headings, ioff, loff)
-
-
-sensor_coordinates.__doc__ += "Notes\n    -----\n" +  schematic
-
-
-def read_xy(csvfile, sep=",", xcol=1, ycol=2, headrows=None):
-
-    try:
-        datafile = open(csvfile, newline="")
-    except TypeError:
-        # handle non-path-like csvfile
-        # csv reader does not require a file object
-        datafile = csvfile
-
-    if headrows is None:
-        headrows_not_passed = True
-        headrows = 0
-    else:
-        headrows_not_passed = False
-    for row in range(headrows):
-        next(datafile)
-
-    reader = iter(
-        csv.reader(datafile, delimiter=sep, quoting=csv.QUOTE_NONNUMERIC)
-    )
-    row_cnt = 0
-    rows_not_yet_read = True
-    xy = []
-    while True:
-        row_cnt += 1
-        try:
-            row = next(reader)
-            rows_not_yet_read = False
-            xy.append((row[xcol - 1], row[ycol - 1]))
-        except StopIteration:
-            break
-        except ValueError:
-            # if errors occur within the first 5 rows, skip as if header
-            if headrows_not_passed and rows_not_yet_read and row_cnt < 6:
-                continue
-            else:
-                raise
-    try:
-        datafile.close()
-    except AttributeError:
-        pass
-
-    return xy
-
-
-def pairwise(iterable):
+def pairwise(s: Iterable) -> Iterator[Tuple[Any, Any]]:
     "s -> (s0,s1), (s1,s2), (s2, s3), ..."
-    a, b = tee(iterable)
+    a, b = tee(s)
     next(b, None)
     return zip(a, b)
 
 
-def platform_headings(gps_coords):
-    """Determine platform heading at each GPS coordinate.
+def headings(points: List[Point]) -> List[StandardPositionAngle]:
+    """Estimate heading at each point in an OTG survey.
 
-    The heading of the sensor platform at each GPS coordinate is
-    estimated as the average bearing of the prior and subsequent legs of
-    the traverse.
+    The heading is computed as the average bearing of the prior and
+    subsequent legs of the traverse.
 
-    In other words, at a given transect point, the platform is assumed
-    to be facing a direction that is the average of the straight-line
-    directions from and towards the previous and next points,
-    respectively.
+    In other words, at each transect point the heading is estimated
+    to be the average of the straight-line directions from and towards
+    the previous and next points, respectively.
     
     The average is weighted by the lengths of the two legs, such that
     shorter legs (nearer transect points) are weighted more heavily.
 
-    Returns
-    -------
-    list of headings given as standard position angles (not bearings)
-
     Notes
     -----
-    Standard Position Angle (SPA) == standard trigonometric unit circle
-    angle having one ray coincident with the positive x-axis, positive
-    in the counter-clockwise direction, and a value between (-pi, pi).
+    Returns headings as standard trigonomic unit circle angles
+    (value between -pi and pi, relative to positive x direction)
+    rather than bearing angles (relative to positive y direction).
 
     """
-    # using SPAs in place of bearings
-    angles = [
-        atan2(y1 - y0, x1 - x0) for (x0, y0), (x1, y1) in pairwise(gps_coords)
-    ]
-    distances = [
-        hypot(x1 - x0, y1 - y0) for (x0, y0), (x1, y1) in pairwise(gps_coords)
-    ]
-    headings = [float("nan")]
+    angles = [atan2(y1 - y0, x1 - x0) for (x0, y0), (x1, y1) in pairwise(points)]
+    distances = [hypot(x1 - x0, y1 - y0) for (x0, y0), (x1, y1) in pairwise(points)]
+    headings = [angles[0]]  # no weighted average for first heading
 
     for (a0, a1), (d0, d1) in zip(pairwise(angles), pairwise(distances)):
         # leg angles are weighted proportional to the other leg's length
@@ -189,34 +84,86 @@ def platform_headings(gps_coords):
         if ave > pi:
             ave -= 2 * pi
         headings.append(ave)
-    headings.append(float("nan"))
+    headings.append(angles[-1])  # no weighted average for last heading
     return headings
 
 
-def offset_coords(gps_coords, headings, inline_offset=0, lateral_offset=0):
-    "Headings must be standard trig angles"
-    gamma = atan2(lateral_offset, inline_offset)
-    rdist = hypot(lateral_offset, inline_offset)
-    coords = [["xgps", "ygps", "xsens", "ysens"]]
-    for (x, y), angle in zip(gps_coords, headings):
-        xsens = x + rdist * cos(angle + gamma)
-        ysens = y + rdist * sin(angle + gamma)
-        coords.append([x, y, xsens, ysens])
-    return coords
+class Survey(List[Point]):
+    """List of coordinates points from OTG survey.
+    
+    Survey = List[Point] = List[Tuple(float, float)]
+    """
+
+    @staticmethod
+    def read_xy(
+        csvfile, sep: str = ",", xcol: int = 1, ycol: int = 2, skiprows: int = 0
+    ) -> Survey:
+        """
+        csvfile : csv filename or object
+            Delimited file containing GPS coordinates for OTG survey.
+        xcol : int, optional
+            Column in datafile with x-coordinates. Default is 1.
+        ycol : int, optional
+            Column in datafile with y-coordinates. Default is 2.
+        sep : str, optional
+            Delimiter used in coordinate datafile. Default is ",".
+        skiprows : int, optional
+            Number of header rows in datafile. Default is 0.
+
+        """
+        try:
+            datafile = open(csvfile, newline="")
+        except TypeError:
+            # handle non-path-like csvfile
+            # csv reader does not require a file object
+            datafile = csvfile
+
+        reader = iter(csv.reader(datafile, delimiter=sep))
+        for _ in range(skiprows):
+            next(reader)
+        xy = []
+        while True:
+            try:
+                row = next(reader)
+                xy.append((float(row[xcol - 1]), float(row[ycol - 1])))
+            except StopIteration:
+                break
+
+        try:
+            datafile.close()
+        except AttributeError:
+            pass
+
+        return Survey(xy)
+
+    def to_sensor_coords(
+        self, inline_offset: int = 0, lateral_offset: int = 0
+    ) -> List[Point]:
+        """
+        inline_offset : float, optional
+            Sensor offset distance inline with the direction of travel.
+            Positive in the direction of travel. Default is 0.
+        lateral_offset : float, optional
+            Sensor offset distance lateral to the direction of travel.
+            Positive to the left (facing forward). Default is 0.
+        """
+        angles = headings(self)
+        gamma = atan2(lateral_offset, inline_offset)
+        radius = hypot(lateral_offset, inline_offset)
+        return [
+            (x + radius * cos(angle + gamma), y + radius * sin(angle + gamma))
+            for (x, y), angle in zip(self, angles)
+        ]
 
 
 def arg_parser():
     parser = argparse.ArgumentParser(
         prog="sensoff",
         formatter_class=argparse.RawDescriptionHelpFormatter,
-        description=(
-            "Correct transect coordinates when sensor is offset from GPS"
-        ),
+        description=("Correct transect coordinates when sensor is offset from GPS"),
         epilog=schematic,
     )
-    parser.add_argument(
-        "FILE", help="Delimited text file with GPS x,y coordinates"
-    )
+    parser.add_argument("FILE", help="Delimited text file with GPS x,y coordinates")
     parser.add_argument(
         "--ioff",
         type=float,
@@ -239,23 +186,21 @@ def arg_parser():
         "--xcol",
         type=int,
         default=1,
-        help="x-coordinate column (default: %(default)d)",
+        help="x-coordinate column (1-based) (default: %(default)d)",
     )
     parser.add_argument(
         "--ycol",
         type=int,
         default=2,
-        help="y-coordinate column (default: %(default)d)",
+        help="y-coordinate column (1-based) (default: %(default)d)",
     )
     parser.add_argument(
-        "--headrows",
+        "--skiprows",
         type=int,
-        help="number of datafile header rows (default: %(default)s)",
+        help="number of datafile header rows to skip (default: %(default)s)",
     )
     parser.add_argument(
-        "--delimiter",
-        default=",",
-        help="datafile delimiter (default: %(default)s)",
+        "--sep", default=",", help="datafile delimiter (default: %(default)s)",
     )
     parser.add_argument(
         "--outfile", default=None, help="write output to file",
@@ -266,17 +211,16 @@ def arg_parser():
 def main():
     parser = arg_parser()
     args = parser.parse_args()
-    coords = sensor_coordinates(
-        gps_file=args.FILE,
-        ioff=args.ioff,
-        loff=args.loff,
+    coords = Survey.read_xy(
+        csvfile=args.FILE,
         xcol=args.xcol,
         ycol=args.ycol,
-        delimiter=args.delimiter,
-        headrows=args.headrows,
-    )
+        sep=args.sep,
+        skiprows=args.skiprows,
+    ).to_sensor_coords(inline_offset=args.ioff, lateral_offset=args.loff)
+
     if args.outfile is None:
-        writer = csv.writer(sys.stdout, delimiter=args.delimiter)
+        writer = csv.writer(sys.stdout, delimiter=args.sep)
         try:
             for row in coords:
                 writer.writerow(row)
@@ -284,7 +228,7 @@ def main():
             pass
     else:
         with open(args.outfile, "w", newline="") as csvfile:
-            writer = csv.writer(csvfile, delimiter=args.delimiter)
+            writer = csv.writer(csvfile, delimiter=args.sep)
             for row in coords:
                 writer.writerow(row)