+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+# v1.1.0 (2021-01-10)
+- Add type hints, re-factor
+- Changed interface from sensor_coords func to Survey class
+- Argument name changes: delimiter -> sep, headerrows -> skiprows 
+- Require specifying skiprows (no longer auto detect header rows)
+
 # v1.0.0 (2020-04-09)
 - Initial release